fix: move to hpp files
