+++ src/DBManager.cpp
@@ -1,4 +1,4 @@
-#include "DBManager.h"
+#include "DBManager.hpp"
 #include <filesystem>
 #include <fstream>
 #include <iostream>
